Sorted out lists being passed through on transactions. [SE]
