+++ quick2wire/parts/test_pcf8591.py
@@ -1,49 +1,9 @@
 
 from quick2wire.i2c_ctypes import I2C_M_RD
 from quick2wire.gpio import In
+from quick2wire.parts.fake_i2c import FakeI2CMaster
 from quick2wire.parts.pcf8591 import PCF8591, FOUR_SINGLE_ENDED, THREE_DIFFERENTIAL, SINGLE_ENDED_AND_DIFFERENTIAL, TWO_DIFFERENTIAL
 import pytest
-
-
-class FakeI2CMaster:
-    def __init__(self):
-        self._requests = []
-        self._responses = []
-        self._next_response = 0
-        self.message_precondition = lambda m: True
-        
-    def all_messages_must(self, p):
-        self.message_precondition
-    
-    def clear(self):
-        self.__init__()
-    
-    def transaction(self, *messages):
-        for m in messages:
-            self.message_precondition(m)
-        
-        self._requests.append(messages)
-        
-        read_count = sum(bool(m.flags & I2C_M_RD) for m in messages)
-        if read_count == 0:
-            return []
-        elif self._next_response < len(self._responses):
-            response = self._responses[self._next_response]
-            self._next_response += 1
-            return response
-        else:
-            return [(0x00,)]*read_count
-    
-    def add_response(self, *messages):
-        self._responses.append(messages)
-    
-    @property
-    def request_count(self):
-        return len(self._requests)
-    
-    def request(self, n):
-        return self._requests[n]
-
 
 i2c = FakeI2CMaster()
 
@@ -55,7 +15,6 @@ def is_write(m):
 
 def assert_is_approx(expected, value, delta=0.005):
     assert abs(value - expected) <= delta
-
 
 def correct_message_for(adc):
     def check(m):
@@ -70,8 +29,6 @@ def correct_message_for(adc):
     
     return check
 
-
-
 def setup_function(f):
     i2c.clear()
 
@@ -83,7 +40,6 @@ def create_pcf8591(*args, **kwargs):
 def assert_all_input_pins_report_direction(adc):
     assert all(adc.single_ended_input(p).direction == In for p in range(adc.single_ended_input_count))
     assert all(adc.differential_input(p).direction == In for p in range(adc.differential_input_count))
-
 
 def test_can_be_created_with_four_single_ended_inputs():
     adc = PCF8591(i2c, FOUR_SINGLE_ENDED)
